Add files via upload
you forgot to add the identifier for the article class, so i did it
"When an article is added to the database, it receives a unique identifier. " so should we overload the Article class so that you can add an article that doesn't have an ID yet but then it gets one? Not sure exactly if that's the idea here.

-we still need a "generateID" method to add IDs to new articles added to the database
-I created all the methods we need in Database, however i'm not sure if I did it right based on concepts in the course and i'm not sure it works, so if you could make a main method that'd be great