Changes part 2
- I added the ID to the @param things
- I added an extra invar saying that the ID should be unique
- An article should not have an ID when you first make it. I think we
need to set it
using a generator function that is also final. Because an ID should not
change.
- Made the ID generator.
- Why did you remove the private keyword from the variables in the
article class?
We make getters & setters for these variables so that they can be
private & encapsulated?
- I fixed the isValidAuthor method. I think it should be called in the
constructor. If you call the constructor by making a new article,
the constructor should say IllegalArgumentException if the given
argument is incorrect. I had to do this in two parts
1) there's a first if clause that checks if the Array of Strings authors
is emtpy. If it is, the exception is shown.
2) there's a second if clause that checks if every string in the array
contains a comma. If this is not the case, the exception is shown.
==> I guess we'll have to do it this way for all the invariants.
- Problem:
database class, line 77: citesA.getCites() => returns null => Java wants
non null object there. So it doens't work.
==> therefore, we are not allowed to initialize these to null, as it
throws a nullpointerexception
- I rewrote the delete function, becaus it threw errors, it can now
delete the articles from the database
- I also changed the way how the cites is updated when an article is
deleted. It works now.